enhancement make rule unique by name and view it is in

+++ be/src/route/v1/POST-update-rules.route.ts
@@ -8,7 +8,7 @@ import {
     vFnHasAnyUserRoles
 } from "./common-middleware";
 import {check, body} from 'express-validator';
-import {doInDbConnection, QueryResponse} from "../../db";
+import {doInDbConnection, QueryA, QueryResponse} from "../../db";
 import {Connection} from "mariadb";
 import {Rule} from "../../model/rule.model";
 import {Rule2} from "../model/server-side.model";
@@ -32,29 +32,47 @@ const httpAction: any[] = [
         const viewId: number = Number(req.params.viewId);
         const rules: Rule[] = req.body.rules;
         const rule2s: Rule2[] = revert(rules);
+        const errors: string[] = [];
         
         for (const rule2 of rule2s) {
             if (rule2.id && rule2.id > 0)  { // update
                 await doInDbConnection(async (conn: Connection) => {
-                    await update(conn, viewId, rule2)
+                    const err: string[] = await update(conn, viewId, rule2)
+                    errors.push(...err);
                 });
             } else { // add
                 await doInDbConnection(async (conn: Connection) => {
-                    await add(conn, viewId, rule2)
+                    const err: string [] = await add(conn, viewId, rule2)
+                    errors.push(...err);
                 });
             }
         }
 
-        res.status(200).json({
-            status: 'SUCCESS',
-            message: `Update successful`
-        } as ApiResponse);
+        if (errors && errors.length) {
+            res.status(200).json({
+                status: 'ERROR',
+                message: errors.join(', ')
+            } as ApiResponse);
+        } else {
+            res.status(200).json({
+                status: 'SUCCESS',
+                message: `Update successful`
+            } as ApiResponse);
+        }
     }
 ];
 
-const add = async (conn: Connection, viewId: number, rule2: Rule2) => {
-
+const add = async (conn: Connection, viewId: number, rule2: Rule2): Promise<string[]> => {
+    const errors: string[] = [];
     await doInDbConnection(async (conn: Connection) => {
+
+            const qq: QueryA = await conn.query(`SELECT COUNT(*) AS COUNT FROM TBL_RULE WHERE NAME=? AND VIEW_ID=?`, [rule2.name, viewId]);
+            if (qq[0].COUNT > 0) { // rule with similar name already exists
+                errors.push(`Rule with name ${rule2.name} already exists in view id ${viewId}`);
+                return;
+            }
+
+
             const rq: QueryResponse = await conn.query(`INSERT INTO TBL_RULE (VIEW_ID, NAME, DESCRIPTION, STATUS) VALUES (?,?,?,'ENABLED')`, [viewId, rule2.name, rule2.description]);
             const ruleId: number = rq.insertId;
 
@@ -88,12 +106,21 @@ const add = async (conn: Connection, viewId: number, rule2: Rule2) => {
                 }
             }
     });
+    return errors;
 };
 
-const update = async (conn: Connection, viewId: number, rule2: Rule2) => {
+const update = async (conn: Connection, viewId: number, rule2: Rule2): Promise<string[]> => {
+    const errors: string[] = [];
     await doInDbConnection(async (conn: Connection) => {
 
         const ruleId: number = rule2.id;
+
+        const qq: QueryA = await conn.query(`SELECT COUNT(*) AS COUNT FROM TBL_RULE WHERE ID=?`, [ruleId]);
+        if (qq[0].COUNT < 0) { // no such rule with id exists, cannot update rule that do not exists
+            errors.push(`Rule with id ${ruleId} do not exists`);
+            return;
+        }
+
         await conn.query(`UPDATE TBL_RULE SET NAME=?, DESCRIPTION=? WHERE ID=?`, [rule2.name, rule2.description, ruleId]);
 
         await conn.query(`DELETE FROM TBL_RULE_VALIDATE_CLAUSE WHERE RULE_ID=?`, [ruleId]);
@@ -131,6 +158,7 @@ const update = async (conn: Connection, viewId: number, rule2: Rule2) => {
             }
         }
     });
+    return errors;
 };
